+++ opencascade-hs/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to the
 
 ## Unreleased
 
+- Add `compose` and `composed` to `OpenCascade.TopoDS.Shape`
 - Add `normalized` and `normalize` to `OpenCascade.GP.Vec` and `OpenCascade.GP.Vec2d`
 
 ## 0.6.2.1